Move the logging util out of contrib, in preparation for use by other Eventing controllers using it.
